2015.7.0

+++ leonardo/module/web/utils/scaffold_web.py
@@ -1,5 +1,10 @@
 
+import logging
+import os
+
+import requests
 import six
+from django.conf import settings
 from django.contrib.auth import authenticate, login
 from django.contrib.auth.models import User
 from horizon_contrib.common import get_class
@@ -7,12 +12,117 @@ from horizon_contrib.common import get_class
 from ..models import (Page, PageColorScheme, PageTheme, WidgetBaseTheme,
                       WidgetContentTheme, WidgetDimension)
 
-from .bootstrap_example import BOOTSTRAP
+LOG = logging.getLogger('leonardo')
+
+LEONARDO_BOOTSTRAP_DIR = getattr(settings, 'LEONARDO_BOOTSTRAP_DIR', None)
 
 
-def create_new_site(run_syncall=False, with_user=True, request=None):
-    """returns page with some widgets
+def _load_from_stream(stream):
+    result = None
+    try:
+        import yaml
+        result = yaml.load(stream)
+    except:
+        pass
+    else:
+        return result
+    try:
+        import json
+        result = json.load(stream)
+    except:
+        pass
+    else:
+        return result
+    return result
+
+
+def get_loaded_scripts(directory=LEONARDO_BOOTSTRAP_DIR):
+    """return dictionary of loaded scripts from specified directory
     """
+
+    scripts = {}
+
+    for root, dirnames, filenames in os.walk(directory):
+
+        for file_name in filenames:
+
+            try:
+
+                ext = file_name.split('.')[1]
+
+                with open(os.path.join(directory, file_name), 'r') as file:
+
+                    scripts[file_name] = _load_from_stream(file)
+
+            except Exception as e:
+                LOG.exception('Error in during loading {} file with {}'.format(
+                    file_name, str(e)))
+
+    return scripts
+
+
+def _handle_regions(regions, feincms_object):
+
+    for region, widgets in six.iteritems(regions):
+        for widget_cls, widget_attrs in six.iteritems(widgets):
+
+            try:
+                WidgetCls = get_class(widget_cls)
+            except Exception as e:
+                raise Exception('Cannout load {} with {}'.format(
+                    widget_cls, e))
+
+            # TODO create form and validate options
+            w_attrs = widget_attrs.get('attrs', {})
+            w_attrs.update({
+                'parent': feincms_object,
+                'region': region,
+                'ordering': 0
+            })
+
+            w_attrs['content_theme'] = WidgetContentTheme.objects.get(
+                name=w_attrs['content_theme'],
+                widget_class=WidgetCls.__name__)
+            w_attrs['base_theme'] = WidgetBaseTheme.objects.get(
+                name=w_attrs['base_theme'])
+            widget = WidgetCls(**w_attrs)
+            widget.save(created=False)
+
+            for size, width in six.iteritems(
+                    widget_attrs.get('dimenssions', {})):
+
+                WidgetDimension(**{
+                    'widget_id': widget.pk,
+                    'widget_type': widget.content_type,
+                    'size': size,
+                    'width': width
+                }).save()
+
+
+def create_new_site(run_syncall=False, with_user=True, request=None,
+                    name='demo.yaml', url=None):
+    """load all available scripts and try scaffold new site from them
+
+    TODO(majklk): refactor and support for more cases
+
+    """
+
+    if run_syncall:
+        from django.core import management
+        management.call_command('sync_all', force=True)
+
+    if url:
+        try:
+            BOOTSTRAP = _load_from_stream(requests.get(url).text)
+        except Exception as e:
+            raise e
+    else:
+        try:
+            scripts = get_loaded_scripts()
+            BOOTSTRAP = scripts[name]
+        except KeyError:
+            raise Exception('Cannot find {} in {} loaded from {}'.format(
+                name, scripts, LEONARDO_BOOTSTRAP_DIR))
 
     root_page = None
 
@@ -36,14 +146,17 @@ def create_new_site(run_syncall=False, with_user=True, request=None):
 
         regions = page_attrs.pop('content', {})
 
+        if not PageTheme.objects.exists() or PageColorScheme.objects.exists():
+            raise Exception("You havent any themes \
+                please install someone and run sync_all")
+
         try:
             if page_theme_name == '__first__':
                 theme = PageTheme.objects.first()
             else:
                 theme = PageTheme.objects.get(name=page_theme_name)
         except Exception:
-            raise Exception("You havent any themes \
-                please install someone and run sync_all")
+            raise Exception(_("You haven't any themes please install someone and run sync_all"))
         else:
             page_attrs['theme'] = theme
 
@@ -64,46 +177,32 @@ def create_new_site(run_syncall=False, with_user=True, request=None):
         # TODO from attrs etc..
         root_page = page
 
-        for region, widgets in six.iteritems(regions):
-            for widget_cls, widget_attrs in six.iteritems(widgets):
-
-                try:
-                    WidgetCls = get_class(widget_cls)
-                except Exception as e:
-                    raise Exception('Cannout load {} with {}'.format(
-                        widget_cls, e))
-
-                # TODO create form and validate options
-                w_attrs = widget_attrs.get('attrs', {})
-                w_attrs.update({
-                    'parent': page,
-                    'region': region,
-                    'ordering': 0
-                })
-
-                w_attrs['content_theme'] = WidgetContentTheme.objects.get(
-                    name=w_attrs['content_theme'],
-                    widget_class=WidgetCls.__name__)
-                w_attrs['base_theme'] = WidgetBaseTheme.objects.get(
-                    name=w_attrs['base_theme'])
-                widget = WidgetCls(**w_attrs)
-                widget.save(created=False)
-
-                for size, width in six.iteritems(
-                        widget_attrs.get('dimenssions', {})):
-
-                    WidgetDimension(**{
-                        'widget_id': widget.pk,
-                        'widget_type': widget.content_type,
-                        'size': size,
-                        'width': width
-                    }).save()
+        _handle_regions(regions, page)
 
     # generic stuff
-    for cls_name, cls_attrs in six.iteritems(BOOTSTRAP):
+    for cls_name, entries in six.iteritems(BOOTSTRAP):
 
-        cls = get_class(cls_name)
+        for entry, cls_attrs in six.iteritems(entries):
 
-        instance, created = cls.objects.get_or_create(**cls_attrs)
+            cls = get_class(cls_name)
+
+            regions = cls_attrs.pop('content', {})
+
+            # load FK from
+            # author: {'pk': 1, 'type': 'auth.User'}
+            for attr, value in six.iteritems(cls_attrs):
+                if isinstance(value, dict):
+                    cls_type = value.get('type', None)
+                    if cls_type:
+                        try:
+                            cls_attrs[attr] = get_class(
+                                cls_type).objects.get(pk=value.get('pk'))
+                        except Exception as e:
+                            raise Exception(
+                                'Cannot load FK {} not Found original exception {}'.format(cls_type, e))
+
+            instance, created = cls.objects.get_or_create(**cls_attrs)
+
+            _handle_regions(regions, instance)
 
     return root_page
